incorrect return type for file_contents strategies
Crystal 0.29.0 resolves the call to `#parse` as a `JSON::Any` type which does not match the return signature of `String` in `file_contents_from_blobs` or `file_contents_from_files`.  This may be because halite only registers `application/json` as a MIME type currently but I'm not sure.  Either way explicitly calling `#to_s` to return the stringified body seems more correct for now at least.